+++ tests/test_user.py
@@ -20,13 +20,17 @@ def test_user_details(user_details, ppms_user):
 def test_user_billing_info(user_details, ppms_connection):
     """Test the PpmsUser billing information."""
     user = PpmsUser(user_details["api_response"], conn=ppms_connection)
-    assert len(user.billing_info) == 2
+    assert len(user.billing_info) == 3
 
     assert user.billing_info[0].billing_code == "pyppms_group_billing_code"
     assert user.billing_info[0].billing_type == "group"
 
     assert user.billing_info[1].billing_code == "pyppms_user_billing_code"
     assert user.billing_info[1].billing_type == "user"
+
+    assert user.billing_info[2].billing_code == "proj.bcode.6"
+    assert user.billing_info[2].billing_type == "project"
+    assert user.billing_info[2].description == "[6] Project Six"
 
 
 def test_user_fullname(user_details, ppms_user):